+++ src/ccs.cpp
@@ -1008,9 +1008,11 @@ namespace detail {
     segment_hash reduced = reduce_all(h, binop).wait();
     if (it != end(segmap)) {
       if (reduced != segment_hash{}) {
-        it->flags |= (uint16_t) segment_flags::verified;
+        it->flags |= static_cast<uint16_t>(segment_flags::verified);
+        it->flags &= ~static_cast<uint16_t>(segment_flags::bad_verification);
       } else {
-        it->flags |= (uint16_t) segment_flags::bad_verification;
+        it->flags |= static_cast<uint16_t>(segment_flags::bad_verification);
+        it->flags &= ~static_cast<uint16_t>(segment_flags::verified);
         throw segment_verification_error("verify_segment() failed: Segment not found on all ranks.");
       }
       flag_map_[it->start] = it->flags;
@@ -1055,10 +1057,13 @@ namespace detail {
       for (auto& seg : segmap) {
         if (seg.ident == hashlist[i])
         {
-          if (checklist[i])
+          if (checklist[i]) {
             seg.flags |= static_cast<uint16_t>(segment_flags::verified);
-          else
+            seg.flags &= ~static_cast<uint16_t>(segment_flags::bad_verification);
+          } else {
             seg.flags |= static_cast<uint16_t>(segment_flags::bad_verification);
+            seg.flags &= ~static_cast<uint16_t>(segment_flags::verified);
+          }
           break;
         }
       }
@@ -1067,8 +1072,10 @@ namespace detail {
     uint16_t found_flags = static_cast<uint16_t>(segment_flags::verified) | static_cast<uint16_t>(segment_flags::bad_verification);
     flag_map_.clear();
     for (auto& seg : segmap) {
-      if (!(seg.flags & found_flags))
+      if (!(seg.flags & found_flags)) {
         seg.flags |= static_cast<uint16_t>(segment_flags::bad_verification);
+        seg.flags &= ~static_cast<uint16_t>(segment_flags::verified);
+      }
       flag_map_[seg.start] = seg.flags;
     }
 #endif